refactor(statement-entries): refactor statement entries which dont use vue filters to vue 3
