+++ components/molecules/statements/DepositEntry.vue
@@ -3,7 +3,11 @@
         <b-col cols="1">
             <div class="transaction-type-icon">
                 <div class="transaction">
-                    <img alt="" src="~/assets/svg/statement/deposit.svg" />
+                    <img
+                        alt="Deposit"
+                        loading="lazy"
+                        src="../../../assets/svg/statement/deposit.svg"
+                    />
                 </div>
             </div>
         </b-col>
@@ -14,28 +18,27 @@
             <div class="text-muted">
                 <b-row>
                     <b-col class="text-right">
-                        <card-fee :transaction="transaction" />
+                        <CardFee :transaction="props.transaction" />
                     </b-col>
                 </b-row>
             </div>
         </b-col>
         <b-col class="text-right" cols="3" xl="2">
-            <amount :transaction="transaction" />
+            <Amount :transaction="props.transaction" />
         </b-col>
     </b-row>
 </template>
-<script lang="ts">
-import { Component, Prop, Vue } from 'nuxt-property-decorator'
+<script lang="ts" setup>
+import { defineProps } from 'vue'
+import { PropType } from '@nuxtjs/composition-api'
+import Amount from '~/components/statement/item/common/Amount.vue'
+import CardFee from '~/components/statement/item/common/CardFee.vue'
 import { StatementEntryModel } from '~/plugins/weavr-multi/api/models/managed-instruments/statements/models/StatementEntryModel'
 
-@Component({
-    components: {
-        Amount: () => import('~/components/statement/item/common/Amount.vue'),
-        CardFee: () => import('~/components/statement/item/common/CardFee.vue'),
+const props = defineProps({
+    transaction: {
+        type: Object as PropType<StatementEntryModel>,
+        required: true,
     },
 })
-export default class StatementItemAdditionalField extends Vue {
-    @Prop()
-    readonly transaction!: StatementEntryModel
-}
 </script>
