+++ components/organisms/StatementItem.vue
@@ -53,7 +53,7 @@ import { PropType } from '@nuxtjs/composition-api'
 import { computed } from 'vue'
 import { StatementEntryModel } from '~/plugins/weavr-multi/api/models/managed-instruments/statements/models/StatementEntryModel'
 import Authorisation from '~/components/statement/item/Authorisation.vue'
-import Deposit from '~/components/statement/item/Deposit.vue'
+import Deposit from '~/components/molecules/statements/DepositEntry.vue'
 import ManualTransaction from '~/components/molecules/statements/ManualTransactionEntry.vue'
 import ChargeFee from '~/components/statement/item/ChargeFee.vue'
 import MerchantRefund from '~/components/molecules/statements/MerchantRefundEntry.vue'
@@ -61,7 +61,7 @@ import Settlement from '~/components/statement/item/Settlement.vue'
 import Transfer from '~/components/molecules/statements/TransferEntry.vue'
 import Send from '~/components/molecules/statements/SendEntry.vue'
 import Withdrawal from '~/components/molecules/statements/WithdrawalEntry.vue'
-import FeeReversal from '~/components/statement/item/FeeReversal.vue'
+import FeeReversal from '~/components/statement/item/FeeReversalEntry.vue'
 import Amount from '~/components/statement/item/common/Amount.vue'
 
 const props = defineProps({
